refactor(deals): the identity line moves out of a file at its ceiling
The frontend length ratchet landed and caught this branch: deals.tsx was
frozen at 4497 and the win-reason fact pushed it to 4553.

The answer the gate asks for, and the one its own seeding script refuses
to let me take instead: shrink rather than re-freeze. The identity line is
a self-contained presentational row with its own type, read by the deal
page and by the 360's tests and stories — so it goes to dealidentity.tsx,
with the value-or-em-dash helper that only it uses, and the win-reason
fact to dealwinreason.tsx beside the stylesheet that clamps it.

deals.tsx: 4497 → 4364.

Not churn for its own sake: the ceiling asks how much a reader must hold
at once, and a screen that has to grow to gain a fact is the case it is
asking about.
